+++ .prettierrc
@@ -1,6 +1,7 @@
 {
   "singleQuote": true,
   "jsxSingleQuote": true,
+  "endOfLine": "auto",
   "semi": false,
   "useTabs": false,
   "tabWidth": 2,
