+++ site.config.js
@@ -1,19 +1,19 @@
 module.exports = {
   // where it all starts -- the site's root Notion page (required)
-  rootNotionPageId: '78fc5a4b88d74b0e824e29407e9f1ec1',
+  rootNotionPageId: '78cd864fb8db44c4b089bf489af4a60a',
 
   // if you want to restrict pages to a single notion workspace (optional)
   // (this should be a Notion ID; see the docs for how to extract this)
   rootNotionSpaceId: null,
 
   // basic site info (required)
-  name: 'Transitive Bullshit',
-  domain: 'transitivebullsh.it',
-  author: 'Travis Fischer',
+  name: 'Ayush Raj',
+  domain: 'ayushcode.com',
+  author: 'Ayush Raj',
 
   // open graph metadata (optional)
   description: 'Example site description',
-  socialImageTitle: 'Transitive Bullshit',
+  socialImageTitle: 'Ayush Raj',
   socialImageSubtitle: 'Hello World! 👋',
 
   // social usernames (optional)
